FInal Submission/requirements.txt
mysql_connector_repackaged==0.3.1
prettytable==3.0.0
